requirement.txt
akracer==0.0.13
akshare==1.11.79
ament-copyright==0.7.12
ament-cppcheck==0.7.12
ament-cpplint==0.7.12
ament-flake8==0.7.12
ament-index-python==0.7.2
ament-lint==0.7.12
ament-lint-cmake==0.7.12
ament-package==0.7.3
ament-pep257==0.7.12
ament-uncrustify==0.7.12
ament-xmllint==0.7.12
appdirs==1.4.4
asttokens==2.4.1
baostock==0.8.8
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
demo-nodes-py==0.7.9
et-xmlfile==1.1.0
examples-rclpy-executors==0.7.5
examples-rclpy-minimal-action-client==0.7.5
examples-rclpy-minimal-action-server==0.7.5
examples-rclpy-minimal-client==0.7.5
examples-rclpy-minimal-publisher==0.7.5
examples-rclpy-minimal-service==0.7.5
examples-rclpy-minimal-subscriber==0.7.5
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.46.0
frozendict==2.3.8
greenlet==3.0.1
html5lib==1.1
idna==3.4
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
jsonpath==0.82.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
launch==0.8.7
launch-ros==0.8.10
lxml==4.9.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
multitasking==0.0.11
mysql-connector==2.2.9
mysqlclient==2.2.0
nest-asyncio==1.5.8
numpy==1.24.4
openpyxl==3.1.2
osrf-pycommon==0.1.9
packaging==23.2
pandas==2.0.3
pandas-datareader==0.10.0
pandas-ta==0.3.14b0
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-mini-racer==0.6.0
Pygments==2.17.2
pyparsing==3.1.1
pypinyin==0.49.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
quality-of-service-demo-py==0.7.9
requests==2.31.0
ros2action==0.7.11
ros2cli==0.7.11
ros2component==0.7.11
ros2launch==0.8.10
ros2lifecycle==0.7.11
ros2msg==0.7.11
ros2multicast==0.7.11
ros2node==0.7.11
ros2param==0.7.11
ros2pkg==0.7.11
ros2run==0.7.11
ros2service==0.7.11
ros2srv==0.7.11
ros2topic==0.7.11
rosidl-runtime-py==0.7.11
rqt-action==0.4.9
rqt-graph==1.0.4
rqt-gui==1.0.7
rqt-gui-py==1.0.7
rqt-reconfigure==1.0.7
rqt-topic==0.4.9
scikit-learn==1.3.2
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
sros2==0.7.2
stack-data==0.6.3
ta-lib-bin @ file:///home/wieneralan/Downloads/ta_lib_bin-0.4.26-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=48b21c0d73d4af4f5b94c1757c6a2179ede07f9c323ce12ad225dd4c6c5199f2
tabulate==0.9.0
teleop-twist-keyboard==2.3.2
threadpoolctl==3.2.0
topic-monitor==0.7.9
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
tushare==1.2.89
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==0.57.0
xlrd==2.0.1
yfinance==0.2.32
zipp==3.17.0
